requirements.txt
Django==2.0.3
mock==2.0.0
mysqlclient==1.3.12
pbr==3.1.1
python-postmark==0.5.0
pytz==2018.3
selenium==3.10.0
six==1.11.0
